connect-four-api/requirements.txt
websockets~=15.0
maturin~=1.8.2
numpy~=2.2.3
opencv-python~=4.11.0.0
torch~=2.6.0
python-snap7~=2.0.2
pytest~=8.3.5
pytest-mock~=3.14.0
pytest-cov~=6.0.0
pygame~=2.6.1
pluggy~=1.5.0
iniconfig~=2.0.0
typing_extensions~=4.12.2
coverage~=7.8.0
